* refactor products ViewModel
