+++ HelloDotNet.Application/Catalog/Products/IPublicProductService.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
 using HelloDotNet.ViewModels.Catalog.Products;
-using HelloDotNet.ViewModels.Catalog.Products.Public;
 using HelloDotNet.ViewModels.Common;
 
 namespace HelloDotNet.Application.Catalog.Products
@@ -9,6 +8,6 @@ namespace HelloDotNet.Application.Catalog.Products
     public interface IPublicProductService
     {
         Task<PagedResult<ProductViewModel>> GetAllByCategoryId(
-            GetProductPagingRequest request);
+            GetPublicProductPagingRequest request);
     }
 }
