+++ HelloDotNet.ViewModels/Catalog/Products/GetManageProductPagingRequest.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using HelloDotNet.ViewModels.Common;
 
-namespace HelloDotNet.ViewModels.Catalog.Products.Manage
+namespace HelloDotNet.ViewModels.Catalog.Products
 {
-    public class GetProductPagingRequest : PagingRequestBase
+    public class GetManageProductPagingRequest : PagingRequestBase
     {
         public string Keyword { get; set; }
         public List<int> CategoryIds { get; set; }
