+++ HelloDotNet.ViewModels/Catalog/Products/GetPublicProductPagingRequest.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using HelloDotNet.ViewModels.Common;
+
+namespace HelloDotNet.ViewModels.Catalog.Products
+{
+    public class GetPublicProductPagingRequest : PagingRequestBase
+    {
+        public int? CategoryId { get; set; }
+    }
+}
