--- HelloDotNet.ViewModels/Catalog/Products/Public/GetProductPagingRequest.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using HelloDotNet.ViewModels.Common;
-
-namespace HelloDotNet.ViewModels.Catalog.Products.Public
-{
-    public class GetProductPagingRequest : PagingRequestBase
-    {
-        public int? CategoryId { get; set; }
-    }
-}
